+++ packages/core/src/index.ts
@@ -21,7 +21,7 @@ import {
 // @ts-ignore untyped library
 import retimer from 'retimer'
 
-import { FULL_VERSION, INTERMEDIATE_HOPS, PACKET_SIZE, VERSION } from './constants.js'
+import { FULL_VERSION, INTERMEDIATE_HOPS, MAX_HOPS, PACKET_SIZE, VERSION } from './constants.js'
 
 import NetworkPeers, { type Entry, NetworkPeersOrigin } from './network/network-peers.js'
 import Heartbeat, { NetworkHealthIndicator } from './network/heartbeat.js'
@@ -54,7 +54,6 @@ import {
   isSecp256k1PeerId,
   type LibP2PHandlerFunction,
   libp2pSendMessage,
-  libp2pSubscribe,
   MIN_NATIVE_BALANCE,
   NativeBalance,
   PublicKey,
@@ -136,24 +135,25 @@ type PeerStoreAddress = {
 
 export type HoprOptions = {
   environment: ResolvedEnvironment
-  announce?: boolean
+  announce: boolean
   dataPath: string
-  createDbIfNotExist?: boolean
-  forceCreateDB?: boolean
-  allowLocalConnections?: boolean
-  allowPrivateConnections?: boolean
-  password?: string
+  createDbIfNotExist: boolean
+  forceCreateDB: boolean
+  allowLocalConnections: boolean
+  allowPrivateConnections: boolean
+  password: string
   connector?: HoprCoreEthereum
-  strategy?: ChannelStrategyInterface
-  hosts?: {
+  strategy: ChannelStrategyInterface
+  hosts: {
     ip4?: NetOptions
     ip6?: NetOptions
   }
-  heartbeatInterval?: number
-  heartbeatThreshold?: number
-  heartbeatVariance?: number
-  networkQualityThreshold?: number
-  onChainConfirmations?: number
+  heartbeatInterval: number
+  heartbeatThreshold: number
+  heartbeatVariance: number
+  networkQualityThreshold: number
+  onChainConfirmations: number
+  checkUnrealizedBalance: boolean
   testing?: {
     // when true, assume that the node is running in an isolated network and does
     // not need any connection to nodes outside of the subnet
@@ -236,8 +236,7 @@ class Hopr extends EventEmitter {
    */
   public constructor(
     private id: PeerId,
-    private db: HoprDB,
-    private connector: HoprCoreEthereum,
+    public db: HoprDB,
     private options: HoprOptions,
     private publicNodesEmitter = new (EventEmitter as new () => HoprConnectConfig['config']['publicNodes'])()
   ) {
@@ -248,7 +247,7 @@ class Hopr extends EventEmitter {
     }
     this.environment = options.environment
     log(`using environment: ${this.environment.id}`)
-    this.indexer = this.connector.indexer // TODO temporary
+    this.indexer = HoprCoreEthereum.getInstance().indexer // TODO temporary
     this.pubKey = PublicKey.fromPeerId(id)
   }
 
@@ -279,7 +278,9 @@ class Hopr extends EventEmitter {
     this.status = 'INITIALIZING'
     log('Starting hopr node...')
 
-    const balance = await this.connector.getNativeBalance(false)
+    const connector = HoprCoreEthereum.getInstance()
+
+    const balance = await connector.getNativeBalance(false)
 
     verbose(
       `Ethereum account ${this.getEthereumAddress().toHex()} has ${balance.toFormattedString()}. Mininum balance is ${new NativeBalance(
@@ -292,7 +293,7 @@ class Hopr extends EventEmitter {
     }
     log('Node has enough to get started, continuing starting payment channels')
     verbose('Starting HoprEthereum, which will trigger the indexer')
-    await this.connector.start()
+    await connector.start()
     verbose('Started HoprEthereum. Waiting for indexer to find connected nodes.')
 
     // Add us as public node if announced
@@ -301,7 +302,7 @@ class Hopr extends EventEmitter {
     }
 
     // Fetch previous announcements from database
-    const initialNodes = await this.connector.waitForPublicNodes()
+    const initialNodes = await connector.waitForPublicNodes()
 
     // Add all initial public nodes to public nodes cache
     initialNodes.forEach((initialNode) => this.knownPublicNodesCache.add(initialNode.id.toString()))
@@ -309,7 +310,7 @@ class Hopr extends EventEmitter {
     // Fetch all nodes that will announces themselves during startup
     const recentlyAnnouncedNodes: PeerStoreAddress[] = []
     const pushToRecentlyAnnouncedNodes = (peer: PeerStoreAddress) => recentlyAnnouncedNodes.push(peer)
-    this.connector.indexer.on('peer', pushToRecentlyAnnouncedNodes)
+    connector.indexer.on('peer', pushToRecentlyAnnouncedNodes)
 
     // Initialize libp2p object and pass configuration
     const libp2p = (await createLibp2pInstance(
@@ -324,13 +325,6 @@ class Hopr extends EventEmitter {
     this.stopLibp2p = libp2p.stop.bind(libp2p)
 
     this.libp2pComponents = libp2p.components
-    // Subscribe to p2p events from libp2p. Wraps our instance of libp2p.
-    const subscribe = (
-      protocols: string | string[],
-      handler: LibP2PHandlerFunction<Promise<Uint8Array> | Promise<void> | void>,
-      includeReply: boolean,
-      errHandler: (err: any) => void
-    ) => libp2pSubscribe(this.libp2pComponents, protocols, handler, errHandler, includeReply)
 
     const sendMessage = ((
       dest: PeerId,
@@ -353,7 +347,7 @@ class Hopr extends EventEmitter {
     )
 
     // react when network registry is enabled / disabled
-    this.connector.indexer.on('network-registry-status-changed', async (enabled: boolean) => {
+    connector.indexer.on('network-registry-status-changed', async (enabled: boolean) => {
       // If Network Registry got enabled, we might need to close existing connections,
       // otherwise there is nothing to do
       if (enabled) {
@@ -371,7 +365,7 @@ class Hopr extends EventEmitter {
     })
 
     // react when an account's eligibility has changed
-    this.connector.indexer.on(
+    connector.indexer.on(
       'network-registry-eligibility-changed',
       async (_account: Address, nodes: PublicKey[], eligible: boolean) => {
         // If account is no longer eligible to register nodes, we might need to close existing connections,
@@ -428,7 +422,7 @@ class Hopr extends EventEmitter {
 
     this.acknowledgements = new AcknowledgementInteraction(
       sendMessage,
-      subscribe,
+      this.libp2pComponents,
       this.getId(),
       this.db,
       (ackChallenge: HalfKeyChallenge) => {
@@ -436,20 +430,21 @@ class Hopr extends EventEmitter {
         this.emit(`hopr:message-acknowledged:${ackChallenge.toHex()}`)
         this.emit('hopr:message-acknowledged', ackChallenge.toHex())
       },
-      (ack: AcknowledgedTicket) => this.connector.emit('ticket:win', ack),
+      (ack: AcknowledgedTicket) => connector.emit('ticket:win', ack),
       () => {},
       this.environment
     )
 
     const onMessage = (msg: Uint8Array) => this.emit('hopr:message', msg)
     this.forward = new PacketForwardInteraction(
-      subscribe,
+      this.libp2pComponents,
       sendMessage,
       this.getId(),
       onMessage,
       this.db,
       this.environment,
-      this.acknowledgements
+      this.acknowledgements,
+      this.options
     )
 
     // Attach socket listener and check availability of entry nodes
@@ -461,15 +456,15 @@ class Hopr extends EventEmitter {
 
     log('libp2p started')
 
-    this.connector.indexer.on('peer', this.onPeerAnnouncement.bind(this))
+    connector.indexer.on('peer', this.onPeerAnnouncement.bind(this))
 
     // Add all entry nodes that were announced during startup
-    this.connector.indexer.off('peer', pushToRecentlyAnnouncedNodes)
+    connector.indexer.off('peer', pushToRecentlyAnnouncedNodes)
     for (const announcedNode of recentlyAnnouncedNodes) {
       await this.onPeerAnnouncement(announcedNode)
     }
 
-    this.connector.indexer.on('channel-waiting-for-commitment', this.onChannelWaitingForCommitment.bind(this))
+    connector.indexer.on('channel-waiting-for-commitment', this.onChannelWaitingForCommitment.bind(this))
 
     try {
       await this.announce(this.options.announce)
@@ -479,7 +474,7 @@ class Hopr extends EventEmitter {
       process.exit(1)
     }
     // subscribe so we can process channel close events
-    this.connector.indexer.on('own-channel-updated', this.onOwnChannelUpdated.bind(this))
+    connector.indexer.on('own-channel-updated', this.onOwnChannelUpdated.bind(this))
 
     this.setChannelStrategy(this.options.strategy || StrategyFactory.getStrategy('passive'))
 
@@ -529,7 +524,7 @@ class Hopr extends EventEmitter {
     if (this.strategy.shouldCommitToChannel(c)) {
       log(`Found channel ${c.getId().toHex()} to us with unset commitment. Setting commitment`)
       try {
-        await retryWithBackoffThenThrow(() => this.connector.commitToChannel(c))
+        await retryWithBackoffThenThrow(() => HoprCoreEthereum.getInstance().commitToChannel(c))
       } catch (err) {
         // @TODO what to do here? E.g. delete channel from db?
         error(
@@ -548,7 +543,7 @@ class Hopr extends EventEmitter {
    */
   private async onOwnChannelUpdated(channel: ChannelEntry): Promise<void> {
     if (channel.status === ChannelStatus.PendingToClose) {
-      await this.strategy.onChannelWillClose(channel, this.connector)
+      await this.strategy.onChannelWillClose(channel)
     }
   }
 
@@ -618,24 +613,32 @@ class Hopr extends EventEmitter {
   }
 
   private async strategyOpenChannel(status: OutgoingChannelStatus) {
-    const destination = peerIdFromString(status.peer_id)
-    const stake = new BN(status.stake_str)
+    try {
+      const destination = peerIdFromString(status.peer_id)
+      const stake = new BN(status.stake_str)
 
-    if (await this.isAllowedAccessToNetwork(destination)) {
-      this.networkPeers.register(destination, NetworkPeersOrigin.STRATEGY_NEW_CHANNEL)
+      if (await this.isAllowedAccessToNetwork(destination)) {
+        this.networkPeers.register(destination, NetworkPeersOrigin.STRATEGY_NEW_CHANNEL)
 
-      const hash = await this.openChannel(destination, stake)
-      verbose('- opened channel', destination, hash)
-      this.emit('hopr:channel:opened', status)
-    } else {
-      error(`Protocol error: strategy wants to open channel to non-registered peer ${destination.toString()}`)
+        const hash = await this.openChannel(destination, stake)
+        verbose('- opened channel', destination, hash)
+        this.emit('hopr:channel:opened', status)
+      } else {
+        error(`Protocol error: strategy wants to open channel to non-registered peer ${destination.toString()}`)
+      }
+    } catch (e) {
+      error(`strategy could not open channel to ${status.peer_id}`, e)
     }
   }
 
   private async strategyCloseChannel(destination: string) {
-    await this.closeChannel(peerIdFromString(destination), 'outgoing')
-    verbose(`closed channel to ${destination.toString()}`)
-    this.emit('hopr:channel:closed', destination)
+    try {
+      await this.closeChannel(peerIdFromString(destination), 'outgoing')
+      verbose(`closed channel to ${destination.toString()}`)
+      this.emit('hopr:channel:closed', destination)
+    } catch (e) {
+      error(`strategy could not close channel ${destination}`)
+    }
   }
 
   private async updateChannelMetrics() {
@@ -663,7 +666,7 @@ class Hopr extends EventEmitter {
       metric_inChannelCount.set(incomingChannels)
       metric_outChannelCount.set(outgoingChannels)
     } catch (e) {
-      log(`error: failed to update channel metrics`, e)
+      error(`error: failed to update channel metrics`, e)
     }
   }
 
@@ -711,7 +714,7 @@ class Hopr extends EventEmitter {
       metric_strategyTicks.increment()
       metric_strategyMaxChannels.set(tickResult.max_auto_channels)
     } catch (e) {
-      log(`failed to do a strategy tick`, e)
+      error(`failed to do a strategy tick`, e)
       throw new Error('error while performing strategy tick')
     }
 
@@ -727,7 +730,7 @@ class Hopr extends EventEmitter {
       await Promise.all(allClosedChannels.map(this.strategyCloseChannel.bind(this)))
       await Promise.all(allOpenedChannels.map(this.strategyOpenChannel.bind(this)))
     } catch (e) {
-      log(`error when strategy was trying to open or close channels`, e)
+      error(`error when strategy was trying to open or close channels`, e)
     }
   }
 
@@ -760,8 +763,7 @@ class Hopr extends EventEmitter {
     this.stopPeriodicCheck?.()
     verbose('Stopping heartbeat & indexer')
     this.heartbeat?.stop()
-    verbose(`Stopping connector`)
-    await this.connector?.stop()
+
     verbose('Stopping database')
     await this.db?.close()
     log(`Database closed.`)
@@ -878,8 +880,9 @@ class Hopr extends EventEmitter {
    * @param msg message to send
    * @param destination PeerId of the destination
    * @param intermediatePath optional set path manually
+   * @param hops optional number of required intermediate nodes
    */
-  public async sendMessage(msg: Uint8Array, destination: PeerId, intermediatePath?: PublicKey[]) {
+  public async sendMessage(msg: Uint8Array, destination: PeerId, intermediatePath?: PublicKey[], hops?: number) {
     if (this.status != 'RUNNING') {
       throw new Error('Cannot send message until the node is running')
     }
@@ -892,7 +895,7 @@ class Hopr extends EventEmitter {
       // Validate the manually specified intermediate path
       await this.validateIntermediatePath(intermediatePath)
     } else {
-      intermediatePath = await this.getIntermediateNodes(PublicKey.fromPeerId(destination))
+      intermediatePath = await this.getIntermediateNodes(PublicKey.fromPeerId(destination), hops)
 
       if (intermediatePath == null || !intermediatePath.length) {
         throw Error(`Failed to find automatic path`)
@@ -1019,7 +1022,7 @@ class Hopr extends EventEmitter {
     const connected = this.networkPeers.debugLog()
 
     let announced: string[] = []
-    for await (const announcement of this.connector.indexer.getAddressesAnnouncedOnChain()) {
+    for await (const announcement of HoprCoreEthereum.getInstance().indexer.getAddressesAnnouncedOnChain()) {
       announced.push(announcement.toString())
     }
 
@@ -1029,10 +1032,10 @@ class Hopr extends EventEmitter {
   }
 
   public subscribeOnConnector(event: string, callback: () => void): void {
-    this.connector.on(event, callback)
+    HoprCoreEthereum.getInstance().on(event, callback)
   }
   public emitOnConnector(event: string): void {
-    this.connector.emit(event)
+    HoprCoreEthereum.getInstance().emit(event)
   }
 
   public startPeriodicStrategyCheck() {
@@ -1072,6 +1075,7 @@ class Hopr extends EventEmitter {
 
     // Address that we will announce soon
     let addrToAnnounce: Multiaddr
+    const connector = HoprCoreEthereum.getInstance()
 
     if (announceRoutableAddress) {
       let multiaddrs = await this.getAddressesAnnouncedToDHT()
@@ -1109,7 +1113,7 @@ class Hopr extends EventEmitter {
     }
 
     // Check if there was a previous annoucement from us
-    const ownAccount = await this.connector.getAccount(this.getEthereumAddress())
+    const ownAccount = await connector.getAccount(this.getEthereumAddress())
 
     // Do not announce if our last is equal to what we intend to announce
     if (ownAccount?.multiAddr?.equals(addrToAnnounce)) {
@@ -1122,7 +1126,7 @@ class Hopr extends EventEmitter {
         'announcing on-chain %s routable address',
         announceRoutableAddress && routableAddressAvailable ? 'with' : 'without'
       )
-      const announceTxHash = await this.connector.announce(addrToAnnounce)
+      const announceTxHash = await connector.announce(addrToAnnounce)
       log('announcing address %s done in tx %s', addrToAnnounce.toString(), announceTxHash)
     } catch (err) {
       log('announcing address %s failed', addrToAnnounce.toString())
@@ -1135,26 +1139,26 @@ class Hopr extends EventEmitter {
     log('setting channel strategy from', this.strategy?.name, 'to', strategy.name)
     this.strategy = strategy
 
-    this.connector.on('ticket:win', async (ack: AcknowledgedTicket) => {
+    HoprCoreEthereum.getInstance().on('ticket:win', async (ack: AcknowledgedTicket) => {
       try {
-        await this.strategy.onWinningTicket(ack, this.connector)
+        await this.strategy.onWinningTicket(ack)
       } catch (err) {
         error(`Strategy error while handling winning ticket`, err)
       }
     })
   }
 
-  public getChannelStrategy(): string {
-    return this.strategy.name
+  public getChannelStrategy(): ChannelStrategyInterface {
+    return this.strategy
   }
 
   public async getBalance(): Promise<Balance> {
-    return await this.connector.getBalance(true)
+    return await HoprCoreEthereum.getInstance().getBalance(true)
   }
 
   public async getNativeBalance(): Promise<NativeBalance> {
     verbose('Requesting native balance from node.')
-    return await this.connector.getNativeBalance(true)
+    return await HoprCoreEthereum.getInstance().getNativeBalance(true)
   }
 
   public smartContractInfo(): {
@@ -1164,7 +1168,7 @@ class Hopr extends EventEmitter {
     hoprNetworkRegistryAddress: string
     channelClosureSecs: number
   } {
-    return this.connector.smartContractInfo()
+    return HoprCoreEthereum.getInstance().smartContractInfo()
   }
 
   /**
@@ -1185,7 +1189,7 @@ class Hopr extends EventEmitter {
     }
 
     const counterpartyPubKey = PublicKey.fromPeerId(counterparty)
-    const myAvailableTokens = await this.connector.getBalance(true)
+    const myAvailableTokens = await HoprCoreEthereum.getInstance().getBalance(true)
 
     // validate 'amountToFund'
     if (amountToFund.lten(0)) {
@@ -1199,7 +1203,7 @@ class Hopr extends EventEmitter {
     }
 
     try {
-      return this.connector.openChannel(counterpartyPubKey, new Balance(amountToFund))
+      return HoprCoreEthereum.getInstance().openChannel(counterpartyPubKey, new Balance(amountToFund))
     } catch (err) {
       this.maybeEmitFundsEmptyEvent(err)
       throw new Error(`Failed to openChannel: ${err}`)
@@ -1213,9 +1217,10 @@ class Hopr extends EventEmitter {
    * @param myFund the amount to fund the channel in my favor HOPR(wei)
    * @param counterpartyFund the amount to fund the channel in counterparty's favor HOPR(wei)
    */
-  public async fundChannel(counterparty: PeerId, myFund: BN, counterpartyFund: BN): Promise<void> {
+  public async fundChannel(counterparty: PeerId, myFund: BN, counterpartyFund: BN): Promise<string> {
+    const connector = HoprCoreEthereum.getInstance()
     const counterpartyPubKey = PublicKey.fromPeerId(counterparty)
-    const myBalance = await this.connector.getBalance(false)
+    const myBalance = await connector.getBalance(false)
     const totalFund = myFund.add(counterpartyFund)
 
     // validate 'amountToFund'
@@ -1230,7 +1235,7 @@ class Hopr extends EventEmitter {
     }
 
     try {
-      await this.connector.fundChannel(counterpartyPubKey, new Balance(myFund), new Balance(counterpartyFund))
+      return connector.fundChannel(counterpartyPubKey, new Balance(myFund), new Balance(counterpartyFund))
     } catch (err) {
       this.maybeEmitFundsEmptyEvent(err)
       throw new Error(`Failed to fundChannel: ${err}`)
@@ -1241,6 +1246,7 @@ class Hopr extends EventEmitter {
     counterparty: PeerId,
     direction: 'incoming' | 'outgoing'
   ): Promise<{ receipt: string; status: ChannelStatus }> {
+    const connector = HoprCoreEthereum.getInstance()
     const counterpartyPubKey = PublicKey.fromPeerId(counterparty)
     const channel =
       direction === 'outgoing'
@@ -1253,20 +1259,30 @@ class Hopr extends EventEmitter {
     }
 
     if (channel.status === ChannelStatus.Open) {
-      await this.strategy.onChannelWillClose(channel, this.connector)
+      await this.strategy.onChannelWillClose(channel)
+    }
+
+    // TODO: should remove this blocker when https://github.com/hoprnet/hoprnet/issues/4194 gets addressed
+    if (direction === 'incoming') {
+      log(
+        `Incoming channel: ignoring closing channel ${channel
+          .getId()
+          .toHex()} because current HoprChannels contract does not support closing incoming channels.`
+      )
+      throw new Error('Incoming channel: Closing incoming channels currently is not supported.')
     }
 
     let txHash: string
     try {
       if (channel.status === ChannelStatus.Open || channel.status == ChannelStatus.WaitingForCommitment) {
         log('initiating closure of channel', channel.getId().toHex())
-        txHash = await this.connector.initializeClosure(counterpartyPubKey)
+        txHash = await connector.initializeClosure(channel.source, channel.destination)
       } else {
         // verify that we passed the closure waiting period to prevent failing
         // on-chain transactions
 
         if (channel.closureTimePassed()) {
-          txHash = await this.connector.finalizeClosure(counterpartyPubKey)
+          txHash = await connector.finalizeClosure(channel.source, channel.destination)
         } else {
           log(
             `ignoring finalizing closure of channel ${channel
@@ -1323,14 +1339,14 @@ class Hopr extends EventEmitter {
   }
 
   public async redeemAllTickets() {
-    await this.connector.redeemAllTickets()
+    await HoprCoreEthereum.getInstance().redeemAllTickets()
   }
 
   public async redeemTicketsInChannel(peerId: PeerId) {
     const selfPubKey = PublicKey.fromPeerId(this.getId())
     const counterpartyPubKey = PublicKey.fromPeerId(peerId)
     const channel = await this.db.getChannelX(counterpartyPubKey, selfPubKey)
-    await this.connector.redeemTicketsInChannel(channel)
+    await HoprCoreEthereum.getInstance().redeemTicketsInChannel(channel)
   }
 
   /**
@@ -1352,11 +1368,11 @@ class Hopr extends EventEmitter {
   }
 
   public async getPublicKeyOf(addr: Address): Promise<PublicKey> {
-    return await this.connector.getPublicKeyOf(addr)
+    return await HoprCoreEthereum.getInstance().getPublicKeyOf(addr)
   }
 
   public async getEntryNodes(): Promise<{ id: PeerId; multiaddrs: Multiaddr[] }[]> {
-    return this.connector.waitForPublicNodes()
+    return HoprCoreEthereum.getInstance().waitForPublicNodes()
   }
 
   // @TODO remove this
@@ -1375,7 +1391,7 @@ class Hopr extends EventEmitter {
   }
 
   public getEthereumAddress(): Address {
-    return this.connector.getPublicKey().toAddress()
+    return HoprCoreEthereum.getInstance().getPublicKey().toAddress()
   }
 
   /**
@@ -1388,7 +1404,7 @@ class Hopr extends EventEmitter {
   public async withdraw(currency: 'NATIVE' | 'HOPR', recipient: string, amount: string): Promise<string> {
     let result: string
     try {
-      result = await this.connector.withdraw(currency, recipient, amount)
+      result = await HoprCoreEthereum.getInstance().withdraw(currency, recipient, amount)
     } catch (err) {
       this.maybeEmitFundsEmptyEvent(err)
       throw new Error(`Failed to withdraw: ${err}`)
@@ -1402,22 +1418,29 @@ class Hopr extends EventEmitter {
    * @returns true if allowed access
    */
   public async isAllowedAccessToNetwork(id: PeerId): Promise<boolean> {
-    return this.connector.isAllowedAccessToNetwork(PublicKey.fromPeerId(id))
+    return HoprCoreEthereum.getInstance().isAllowedAccessToNetwork(PublicKey.fromPeerId(id))
   }
 
   /**
-   * Takes a destination and samples randomly intermediate nodes
+   * Takes a destination, and optionally the desired number of hops,
+   * and samples randomly intermediate nodes
    * that will relay that message before it reaches its destination.
    *
    * @param destination instance of peerInfo that contains the peerId of the destination
+   * @param hops optional number of required intermediate nodes (must be an integer 1,2,...MAX_HOPS inclusive)
    */
-  private async getIntermediateNodes(destination: PublicKey): Promise<PublicKey[]> {
+  private async getIntermediateNodes(destination: PublicKey, hops?: number): Promise<PublicKey[]> {
+    if (!hops) {
+      hops = INTERMEDIATE_HOPS
+    } else if (![...Array(MAX_HOPS).keys()].map((i) => i + 1).includes(hops)) {
+      throw new Error(`the number of intermediate nodes must be an integer between 1 and ${MAX_HOPS} inclusive`)
+    }
     return await findPath(
       PublicKey.fromPeerId(this.getId()),
       destination,
-      INTERMEDIATE_HOPS,
+      hops,
       (p: PublicKey) => this.networkPeers.qualityOf(p.toPeerId()),
-      this.connector.getOpenChannelsFrom.bind(this.connector)
+      HoprCoreEthereum.getInstance().getOpenChannelsFrom.bind(HoprCoreEthereum.getInstance())
     )
   }
 
@@ -1437,7 +1460,7 @@ class Hopr extends EventEmitter {
             try {
               // call connector directly and don't use cache, since this is
               // most likely outdated during node startup
-              const nativeBalance = await this.connector.getNativeBalance(false)
+              const nativeBalance = await HoprCoreEthereum.getInstance().getNativeBalance(false)
               if (nativeBalance.toBN().gte(MIN_NATIVE_BALANCE)) {
                 resolve()
               } else {
@@ -1465,6 +1488,7 @@ class Hopr extends EventEmitter {
       )
       // Close DB etc.
       await this.stop()
+      await HoprCoreEthereum.getInstance().stop()
       process.exit(1)
     }
   }